lab04 uploaded
There was a merging error with the allocator.c - ive fixed it, and added
my edits. Just forgot to add comments. I included a check for if already
initialised, and optimised the round to nearest power of n function
(alhough could still be better). Rather than delete your code, I've
tried to comment it out and tab it far out of the file, so when we meet
we can safely delete it