Unable to log into UNSW servers
So awesomely, I'm unable to log into the servers right now, and work on
the code!

I've just split the file so that the one you uploaded (full of comments)
is appended with 338am..., and graph.c has my code deleted and yours
uncommented.

I'm out a good chunk of the day but if the connection starts to wrok
again, I'll have a crack