It works 90%!
I've submitted on OL - have a look at the program, I've adapted it from
the GraphLab files and it currently has two problems. 1. It prints the
list in reverse order (need to change how bestPath is generated, may
need another temporary array. 2. It definitely generates a path, and
changing the max distance does alter the mechanics, but the paths are
usually much longer than one would expect (and the examples on the lab
page). I think that it may be a problem when comparing the distances to
the max (might need to convert from KM to Miles).

If you could fix those two issues before the lab on Wednesday I'd really
appreciate it. I think we were both way off in our original programs,
but for this I've just gone line by line and adapted it to our program.
Link to the submission:

https://www.openlearning.com/content/54159e03f860612e8c9c5b60